modified:   bluez/declarativemanager.cpp 	modified:   screen/screen.cpp

+++ bluez/declarativemanager.cpp
@@ -30,7 +30,7 @@ static DeclarativeAdapter *adaptersAtFunction(QQmlListProperty<DeclarativeAdapte
     return manager->m_adapters.values().at(index);
 }
 
-static int devicesCountFunction(QQmlListProperty<DeclarativeDevice> *property)
+static long long int devicesCountFunction(QQmlListProperty<DeclarativeDevice> *property)
 {
     Q_ASSERT(qobject_cast<DeclarativeManager *>(property->object));
     DeclarativeManager *manager = static_cast<DeclarativeManager *>(property->object);
@@ -38,7 +38,7 @@ static int devicesCountFunction(QQmlListProperty<DeclarativeDevice> *property)
     return manager->m_devices.count();
 }
 
-static DeclarativeDevice *devicesAtFunction(QQmlListProperty<DeclarativeDevice> *property, int index)
+static DeclarativeDevice *devicesAtFunction(QQmlListProperty<DeclarativeDevice> *property, long long int index)
 {
     Q_ASSERT(qobject_cast<DeclarativeManager *>(property->object));
     DeclarativeManager *manager = static_cast<DeclarativeManager *>(property->object);
